requirements.txt
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.0.0
Pillow>=9.0.0
django-filter>=23.0
python-decouple>=3.6
psycopg2-binary>=2.9.0
redis>=4.0.0
django-redis>=5.0.0
celery>=5.2.0
django-storages>=1.13.0
boto3>=1.26.0
argon2-cffi>=21.3.0
django-ratelimit>=4.0.0
